+++ misc/pstree.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+# echo $BASHPID > /tmp/pid
+  pstree -U --hide-threads -c -a `cat /tmp/pid`